requirements.txt
asgiref==3.7.2
Django==4.2.2
djangorestframework==3.14.0
gunicorn==20.1.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.0
whitenoise==6.5.0
